requirements.txt
altair==5.5.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
feedparser==6.0.12
fonttools==4.60.1
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
holidays==0.82
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.8.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
pkg==0.2
plotly==6.3.1
prophet==1.1.7
protobuf==6.33.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stanio==0.5.1
statsmodels==0.14.5
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
